Return original error message when Transactional fails
This commit modifies the Transactional method to return the original
error message that caused the transaction function to fail. Previous
to this change, if the transaction function failed, Transactional would
call tx.Rollback() and return the error from that, which is often nil.

With this change, the original error is returned. If the rollback fails
as well, an appropriate indication will be added to the original error
message as well.