Add support for writing conditions directly in SQL
This commits adds a new condition structure (implementing the
WhereCondition interface) called SQLCondition. These are conditions
written directly in SQL, meant to be used for writing complex SQL
conditions or other conditions that are not currently supported by sqlz.

Sample usage:

    z.
        Select("*").
        From("table").
        Where(
            sqlz.Eq("name", "Some Name"),
            sqlz.SQLCond(
                "? LIKE translate(some_col, '*', '%')",
                "text",
            ),
        )

This query will result in the following SQL:

    SELECT *
      FROM table
     WHERE name = ? AND ? LIKE translate(some_col, '*', '%')

Note that question marks must be used as placeholders regardless of the
database driver used, as sqlz automatically converts placeholders to the
appropriate format.