Add support for locks in SELECT statements
This commit adds the ability to lock rows and tables when issuing SELECT
queries. The FOR UPDATE, FOR NO KEY UPDATE, FOR SHARE and FOR KEY SHARE
lock strengths are implemented, with the ability to confine them two
specific tables and to not wait if locks cannot be acquired. Multiple
locks can be performed within one SELECT statement.

Sample code:

    var data someStruct
    err = db.
        Select("*").
        From("table").
        InnerJoin(
            "foreign",
            sqlz.Eq("foreign.id", sqlz.Indirect("table.id"),
        ).
        Where(sqlz.Eq("id", 1)).
        Lock(sqlz.ForUpdate().OfTables("table").SkipLocked()).
        GetRow(&data)