Add SetIf to UPDATE and ON CONFLICT
This commit adds the SetIf() method to both the UpdateStmt and
ConflictClause struct, representing UPDATE statements and ON CONFLICT
clauses, respectively. SetIf works exactly like Set, except that it also
accepts a boolean value, and only does anything if that value is true.

The reasoning behind this method is to allow conditional updates in a
statements without having to save the statement in a separate variable
and using if statements. For example, instead of this:

    conflict := z.OnConflict("email").
                DoUpdate().
                Set("date", time.Now())

    if someVariable > 3 {
        conflict.Set("some-col", "some-val")
    }

    z.
        InsertInto("last_logins").
        Columns("email", "date").
        Values("user@domain.tld", time.Now()).
        OnConflict(conflict).
        Exec()

We can do this:

    z.
        InsertInto("last_logins").
        Columns("email", "date").
        Values("user@domain.tld", time.Now()).
        OnConflict(
            z.OnConflict("email").
                DoUpdate().
                Set("date", time.Now()).
                SetIf("some-col", "some-val", someVariable > 3),
        ).
        Exec()