Add ability to receive results as maps
This commit adds the ability to receive the results from a SELECT
statement as map[string]interface{} (when getting one row of results) or
a slice of such maps (if getting multiple rows).

As opposed to GetRow() and GetAll(), there is no need to create a
variable beforehand; sqlz returns the maps by itself.

Sample usage:

    asMap, err := z.
        Select("string_col", "int_col").
        From("some_table").
        Where(sqlz.Eq("id", 3)).
        GetRowAsMap()

    if asMap["int_col"].(int) == 13 {
        println("I'm bad at writing example code");
    }