Add support for NOT operator and LIKE ANY
This commit adds a pre-condition type that allows attaching operators as condition prefixes.
The 'NOT' operator is implemented and can be used like this:

	sqlz.Not(sqlz.Eq("some_column", sqlz.Indirect("another_column")))

This commit also adds support of the key words 'LIKE ANY', which allows checking for patterns in an array,
whether the array input is a column or an actual array type.
the implementation includes changing the parsing of all arrayCondition types,
which now accepts not only array columns but also array types.

'LIKE ANY' could be used like this:

	sqlz.LikeAny(some_array, sqlz.Indirect("some_column"))

or like this:

	sqlz.LikeAny("some_array_column", sqlz.Indirect("some_column"))