Support GetCount on union queries
This commit fixes GetCount on union statements, and update
columns/limits/offsets on all the union statements instead of applying
it on the first statement only.
There was a runtime error on a union query, because the columns of the union
statements remained as they were, while the columns of the first statement
changed.