Add support to Set command.
The SET command changes run-time configuration parameters of PostgreSQL.
This commit adds support to the SET command. User may also run the SET
command with the LOCAL or SESSION annotation options to control the
lifespan of the coniguration.

One common use case of the Set command is to set statement timeout;
For that reason: a convenient method 'SetTimeout' was added to the
transaction struct.